+++ src/routes/application/my.js
@@ -15,7 +15,7 @@ var vars = require("../../variables");
 
 const ApplicationTable = memo(({ showDetail }) => {
     const { t: tr } = useTranslation();
-    const { userSettings, applicationTypes } = useContext(AppContext);
+    const { userSettings, applicationTypes, loadApplicationTypes } = useContext(AppContext);
 
     const columns = [
         { id: 'id', label: 'ID', orderKey: 'applicationid', defaultOrder: 'desc' },
@@ -46,6 +46,11 @@ const ApplicationTable = memo(({ showDetail }) => {
         async function doLoad() {
             window.loading += 1;
 
+            let localApplicationTypes = applicationTypes;
+            if (applicationTypes === null) {
+                localApplicationTypes = await loadApplicationTypes();
+            }
+
             let processedParam = removeNUEValues(listParam);
 
             let [_recent, _applications] = [{}, {}];
@@ -65,7 +70,7 @@ const ApplicationTable = memo(({ showDetail }) => {
             let newApplications = [];
             for (let i = 0; i < _applications.list.length; i++) {
                 let app = _applications.list[i];
-                newApplications.push({ id: app.applicationid, type: applicationTypes[app.type]?.name ?? tr("unknown"), submit: <TimeAgo key={`${+new Date()}`} timestamp={app.submit_timestamp * 1000} />, update: <TimeAgo key={`${+new Date()}`} timestamp={app.respond_timestamp * 1000} />, staff: <UserCard user={app.last_respond_staff} />, status: STATUS[app.status], application: app });
+                newApplications.push({ id: app.applicationid, type: localApplicationTypes ? (localApplicationTypes[app.type]?.name ?? tr("unknown")) : tr("unknown"), submit: <TimeAgo key={`${+new Date()}`} timestamp={app.submit_timestamp * 1000} />, update: <TimeAgo key={`${+new Date()}`} timestamp={app.respond_timestamp * 1000} />, staff: <UserCard user={app.last_respond_staff} />, status: STATUS[app.status], application: app });
             }
 
             if (pageRef.current === page) {
@@ -76,7 +81,7 @@ const ApplicationTable = memo(({ showDetail }) => {
             window.loading -= 1;
         }
         doLoad();
-    }, [page, pageSize, STATUS, listParam]);
+    }, [page, pageSize, STATUS, listParam]); // do not include applicationTypes to prevent rerender loop on network error
 
     function handleClick(data) {
         showDetail(data.application);
@@ -87,7 +92,7 @@ const ApplicationTable = memo(({ showDetail }) => {
             <Grid item xs={12} sm={12} md={recent.length === 2 ? 6 : 12} lg={recent.length === 2 ? 6 : 12}>
                 <Card>
                     <CardContent>
-                        <Typography variant="subtitle2" gutterBottom>{tr("recent")} {applicationTypes[recent[0].type].name}{tr("application")}</Typography>
+                        <Typography variant="subtitle2" gutterBottom>{tr("recent")} {applicationTypes !== null ? applicationTypes[recent[0].type].name : tr("unknown")} {tr("application")}</Typography>
                         <Typography variant="h5" component="div">
                             {STATUS[recent[0].status]}
                         </Typography>
@@ -101,7 +106,7 @@ const ApplicationTable = memo(({ showDetail }) => {
                 <Grid item xs={12} sm={12} md={6} lg={6}>
                     <Card>
                         <CardContent>
-                            <Typography variant="subtitle2" gutterBottom>{tr("recent")} {applicationTypes[recent[1].type].name}{tr("application")}</Typography>
+                            <Typography variant="subtitle2" gutterBottom>{tr("recent")} {applicationTypes !== null ? applicationTypes[recent[1].type].name : tr("unknown")} {tr("application")}</Typography>
                             <Typography variant="h5" component="div">
                                 {STATUS[recent[1].status]}
                             </Typography>